+++ libs/libcommon/pyproject.toml
@@ -28,7 +28,6 @@ pytz = "^2020.1"
 soundfile = ">=0.12.1"
 starlette-prometheus = "^0.9.0"
 tqdm = "^4.66.3"
-polars = "0.20.15"
 
 [tool.poetry.group.dev.dependencies]
 aiobotocore = "^2.7.0"
